Min stable

+++ mmdet/core/evaluation/eval_mr.py
@@ -482,7 +482,7 @@ class Params:
         self.catIds = []
         # np.arange causes trouble.  the data point on arange is slightly larger than the true value
 
-        self.recThrs = np.linspace(.0, 1.00, round((1.00 - .0) / .01) + 1, endpoint=True)
+        self.recThrs = np.linspace(.0, 1.00, int(round((1.00 - .0) / .01) + 1), endpoint=True)
         self.fppiThrs = np.array([0.0100,    0.0178,    0.0316,    0.0562,    0.1000,    0.1778,    0.3162,    0.5623,    1.0000])
         self.maxDets = [1000]
         self.expFilter = 1.25
